Moved creation of a new Git repo to the end of the checklist
The creation of a new Git repo is now moved to the end of the checklist. It
was not in place as the third step before. Now, the user makes all the
standard changes and just before they start going into the details of their
research, they define their new repository.