Basic titles now corrected after copyright in README.md
The basic title, short description and author section of the checklist was
moved immediately after copyright because it is more related to it. The
description of Gnuastro is now before `README' because the user will
confront Gnuastro in `README' and this will provide a good introduction on
what to do with it.