Corrected tips title to be more clear
The title "Tips on using the pipeline" was a little generic and could be
confused with people who want to reproduce the result, not the designers of
the pipeline. So it was changed to "Tips on expanding this template
(designing your pipeline)". Some minor edits were also made to its first
paragraph.