Choice to build final PDF removed from LOCAL settings
The previous change where we had set the building of the PDF as a local
(and thus not version controlled) setting was not good, because different
commits might be made without the high-level preparations for the final PDF
(especially during the initial/testing phases of a research). Therefore, if
the runner of the pipeline is ignorant to this, they may hit some errors in
LaTeX which can be frustrating.

To have a clean reproduction, it is thus necessary to have the choice of
pdf-building under version control along with the rest of the pipeline.