Configure script starts with bin/bash shebang
While trying the pipeline on a remote server (which runs on Debian), the
configure script had an `Syntax error: "(" unexpected' error. This is
caused by the fact that in the Debian world (and its derivate OSs), the
default shell is not Bash but Dash which has much fewer features for fast
loading. It was thus necessary to start the configure script explicity with
the `/bin/bash' shebang.