Added tips section on version control with tagging
In the "Tips" section of `README.md', a section on version control was
added and a first tip regarding tags was added to guide (new) users on how
to effectively define and use tags. As a result, in the checklist, adding a
`v0' tag is now also suggested.

Some minor typos were also fixed.