Added link to slides on this concept
I had made some slides for a talk here at Lyon Observatory a little over a
week ago and I thought it may be useful to add them in the `README.md' file
to help demonstrate the general concept before having to read such long
texts.

Later, we should be adding some figures to this `README.md' file to make it
more easier to understand.