Copyrights and TeX management made more clear
Until now, the copyright statement was left empty for the users of the
pipeline to fill. However, the files have already been created and have an
author (or contributing authors) before the user starts using the
pipeline. So the original authors of the files are added along with the
year. The user can add their own name to the existing files under the
"Contributing author" when they start and they will be the "Original
author" of the new files they create.

Several changes were also made to the TeX management:

 - LaTeX is run within a `reproduce/build/tex/build' directory now. Not in
   the top reproduction pipeline directory. This helps keep all the
   auxiliary TeX files and directories in that directory and keep the top
   reproduction pipeline directory clean. After the final PDF is built, a
   copy is put in the top reproduction pipeline directory for easy viewing.

 - The PGFPlots preamble was also made more useful, allowing the name of
   the `.tex' file to also be the name of the final plot that is
   produced. This is a GREAT feature, because without it, the TiKZ
   externalization would be based on order of the plots within the
   paper. But now, order is irrelevant and we can even delete the TiKZ
   files within the processing workhorse-Makefiles so the plots are
   definitly rebuilt on the next run.

 - The paper is now in a two-column format to be more similar to published
   papers.

A tip on debugging Make was added to `README.md'.