SHELL has been explicitly set to /bin/bash
On some systems, the default shell `/bin/sh' doesn't point to Bash and
this can cause problems and failures when the designer uses its
features. Bash (and its extra features) make things very easy and it
is very ubiquitous, so it is safe to assume users will have it.

This problem was reported by Alejandro Serrano Borlaff.