Dependencies built at the start of the pipeline
To enable easy/proper reproduction of results, all the high-level
dependencies are now built within the pipeline and installed in a fixed
directory that is added to the PATH of the Makefile. This includes GNU Bash
and GNU Make, which are then used to run the pipeline.

The `./configure' script will first build Bash and Make within itself, then
it will build

All the dependencies are also built to be static. So after they are built,
changing of the system's low-level libraries (like C library) won't change
the tarballs.

Currently the C library and C compiler aren't built within the pipeline,
but we'll hopefully add them to the build process also.

With this change, we now have full control of the shell and Make that will
be used in the pipeline, so we can safely remove some of the generalities
we had before.