Most library versions are now also checked
All the libraries that define their version string as a macro in their
headers are now also checked in `reproduce/src/make/initialize.mk'.

Also, the CFITSIO tarball now follows the same versioning style as the rest
of the tarballs: a script is added to convert the version string into what
is included in the tarball.