File is made before calling flock (file locking)
While testing on another computer, I noticed that to operate properly, the
file given to `flock' must be created before it is called. This is a
low-level difference (how the system treats files), so it wasn't apparent
on my system. To fix it, we have added a `touch' command before it.