Preference for shared library linking
Some high-level programs like Wget and cURL need to be built in shared mode
because they also include dynamic loading of libraries. Therefore, if we
only build the lower-level libraries in static mode, our own build will be
ignored and they will go and find the system's shared libraries to link
with. Because of this, for now, we have manually set the `static_build'
variable in the configure script to `no'.

Also, if the downloader fails, we'll delete the output (an empty file in
the case of Wget) because it interefers with a target definition.