Low-level links nolonger a prerequisite of programs
Until now the low-level links that we put in our internal installation from
the operating system were a prerequisite of essentially all the basic
dependencies. So a change in them would mean a full re-build of all the
basic dependencies. But in building the basic dependencies, we already have
the operating system's PATH and other environment variables. So unlike the
higher-level dependencies, they don't need these links at all!

With this commit, the `low-level-links' file is placed in `installed/bin'
and is a top-level target of the basic dependencies build. In this way, if
it is necessary to update/change to use something from the host operating
system, we can simply delete it and run `./configure' again wihout having
to re-build all the basic dependencies.