Removed already built symbolic links to OS tools
`ln' will complain about a link already existing. So to avoid having to
rely on the `-f' option (which may not be portable across systems), when we
are making symbolic links to the OS tools that we won't be building, we now
remove the file if it exists, then make a new symbolic link.