Shared library absolute address fixed in Libgit2 and WCSLIB on Mac OS
The build systems of Libgit2 and WCSLIB on Mac OS does not account for
installation in non-standard addresses: `Libgit2' keeps the absolute
address of its build directory (not the installation directory) and WCSLIB
doesn't write any absolute address at all (so the system uses the first one
it finds).

To address these issues, we are now using Mac OS's `install_name_tool'
program to fix the absolute path within the installed shared library.

Since the version of the library is actually present in its shared library
name, in `dependency-versions.mk' we have also separated these two
libraries so later when their version is changed, we are careful in
correcting the shared library name also.