Forced linking of bin/sh and verbose Make in Git
While checking the build of the previous commit, a failure happened when
linking `reproduce/build/dependencies/installed/bin/sh' with the built Bash
(because the symbolic link already existed!). So a `-f' flag was added to
`ln' to just change it without complaining.

I also noticed that the Git build was also not in verbose mode. So this has
also been corrected.