Updated Gzip and Gnuastro versions to standard builds
Both Gzip and Gnuastro were being bootstrapped personally from their Git
repository until now. But fortunately a new release of both came out last
week and so to make things standard we are now using their standard
tarballs.

I also noticed that we weren't checking the version of Gzip or mentioning
it in the acknowledgement section. This was also corrected.