Links to the built libncursesw corrected after its build
In a previous implementation, we were using a `target' variable to define
the final target of several links, but with the new `sov' solution, we just
used its base name. However, we had forgot to correct two instances of
`target'. This is corrected now.

Also, the step to clean all already built outputs of the NCURSES library
has been simplified to a platform independent wildcard.